refactor: 다이어리 폼 및 액션 타입 최적화
- ServerDiary 타입 정의를 actions.ts로 이동
- DiaryForm에서 diaryDate 관련 상태 및 로직 제거
- TiptapEditor 이미지 삽입 로직 간소화
- 불필요한 타입 및 상태 제거로 코드 정리
