+++ src/app/actions.ts
@@ -1,7 +1,14 @@
 'use server';
 
 import { revalidatePath } from 'next/cache';
-import { updateDiary, deleteDiary, createDiary, ServerDiary } from '@/lib/diary';
+import { updateDiary, deleteDiary, createDiary, Diary } from '@/lib/diary';
+
+export type ServerDiary = Partial<Omit<Diary, 'content'>> & {
+  content: string;
+  title: string;
+  emotion: string;
+  weather: string;
+};
 
 const sanitizeContent = (content: Record<string, any>) => {
   return JSON.parse(JSON.stringify(content));
@@ -30,6 +37,8 @@ export async function createDiaryAction(data: ServerDiary) {
 
 // 일기 수정 서버 액션
 export async function updateDiaryAction(id: string, data: ServerDiary) {
+  const content = typeof data.content === 'string' ? JSON.parse(data.content) : data.content;
+  data.content = sanitizeContent(content);
   const result = await updateDiary(id, data);
 
   if (result.error) {
